Project: It now draws a triangle!
Summary:
I have setup most of the basic OpenGL constructs to set up any simple vertex
draws like a triangle. This commit shows the basics in loading vertex buffers,
vertex array objects and loading shader source objects and shader programs.

I have also added a couple of utility functions to make life easier such as a
file contents loading function. I've decided to keep the shader files at the
root directory at the project, at the cost that running program must be
executed in the root directory as well, in order to find those shader or
other data files.

In the future, it may be wise to create a procedure
to import both the program and the data files in the same directory if the
program is to be distributed somewhere else.